requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==4.0.5
django-environ==0.8.1
django-heroku==0.3.1
django-tinymce==3.4.0
gunicorn==20.1.0
idna==3.3
mailchimp-marketing==3.0.75
Pillow==9.1.1
psycopg2==2.9.3
python-dateutil==2.8.2
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.2.0
